Move aggregation tests to separate file
